requirements.unstable.txt
asgiref
bleach
boto3
channels
channels-redis
cryptography>=2.3
csscompressor
django-annoying
django-anymail[mailgun]
django-compressor
django-celery-beat
django-celery-results
django-cors-headers
django-debug-toolbar
django-extensions
django-heroku-postgresify
django-hosts
django-htmlmin
django-ipware
django-localflavor
django-redis-cache
django-storages
django-static-precompiler[watch]
Django
extraction
Faker
flake8
freezegun
gunicorn
hashids
honcho
html5lib
ipython
kombu>=3.0.30
jwt
lxml
mistune
mock
Pillow
piprot
polytester>=1.2.0
psycopg2
pytz
python-magic
PyYAML>=4.2b1
stripe
sorl-thumbnail
tblib
user_agents
uvicorn
